app2/requirements.txt
flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp
opentelemetry-exporter-prometheus
prometheus-client

